extract shared functions into a single file
